+++ src/graph/nodes/InputNode.cpp
@@ -30,7 +30,7 @@ namespace arm_compute
 {
 namespace graph
 {
-InputNode::InputNode(TensorDescriptor desc, unsigned int size) : _desc(std::move(desc))
+InputNode::InputNode(TensorDescriptor desc, size_t size) : _desc(std::move(desc))
 {
     _outputs.resize(size, NullTensorID);
 }
